refactor: Move OptimizedPredictionPipeline class to new file, move short_label to utils.py, update realtime for connections
