+++ utils.py
@@ -58,6 +58,28 @@ class ConfigError(EEGSystemError):
     """Raised when configuration is invalid or missing."""
 
 
+# Mapping for short label display (for direction labels)
+SHORT_LABELS = {
+    "forward": "FWD",
+    "backward": "BWD",
+    "left": "LFT",
+    "right": "RGT",
+    "neutral": "NEU",
+}
+
+def short_label(label: str) -> str:
+    """Return a short display label for a given direction label.
+
+    Args:
+        label (str): The full direction label.
+
+    Returns:
+        str: The short label (e.g., "FWD" for "forward").
+
+    """
+    return SHORT_LABELS.get(label.lower(), label[:3].upper())
+
+
 def setup_logging(
     logfile: str = "eeg_training.log", default_level: str | None = None,
 ) -> None:
